Extract `DatabaseTableView` component
